Feat: Add App and input events catcher and  moved
